+++ peers/index.js
@@ -2,12 +2,10 @@ const findTagMatch = require('./find_tag_match');
 const interceptInboundChannels = require('./intercept_inbound_channels');
 const limitForwarding = require('./limit_forwarding');
 const openChannels = require('./open_channels');
-const openChannelsForSavedNodes = require('./open_channels_for_saved_nodes');
 
 module.exports = {
   findTagMatch,
   interceptInboundChannels,
   limitForwarding,
   openChannels,
-  openChannelsForSavedNodes,
 };
